add doubly linked list to tutorial
