+++ src/examples/Dbl_Linked_List/c_types.h
@@ -3,8 +3,3 @@ struct node {
   struct node* prev;
   struct node* next;
 };
-
-struct node_and_int {
-  struct node* node;
-  int data;
-};
